Create Best Coupon
Chef is ordering food online (instead of cooking) and the bill comes out to be Rs. XX. Chef can use one of the following two coupons to avail a discount.

Get 1010 percent off on the bill amount
Get a flat discount of Rs. 100100 on the bill amount
What is the maximum discount Chef can avail?